User/requirements.txt
streamlit
langchain
faiss-cpu
boto3
langchain_community
python-dotenv
streamlit-elements
pypdf